+++ app/javascript/applications/HomebrewsApp/pages/Content/Cosmere/Settings.jsx
@@ -1,0 +1,31 @@
+import { useAppState } from '../../../context';
+import { SharedContent } from '../../../pages';
+import { fetchListRequest, fetchHomebrewRequest, batchDestroyRequest } from '../../../requests_v2/list';
+import { removeSettingRequest } from '../../../requests_v2/cosmere/settings';
+
+export const CosmereSettings = () => {
+  const [appState] = useAppState();
+
+  const fetchList = async () => await fetchListRequest(appState.accessToken, 'Cosmere::Homebrews::Setting');
+  const fetchHomebrew = async (id) => await fetchHomebrewRequest(appState.accessToken, 'Cosmere::Homebrews::Setting', id);
+  const batchDestroy = async (ids) => await batchDestroyRequest(appState.accessToken, 'Cosmere::Homebrews::Setting', ids);
+
+  const ChildrenComponent = (props) => (
+    <div class="flex flex-col gap-4">
+      <p>ID - {props.info.id}</p>
+    </div>
+  );
+
+  return (
+    <SharedContent
+      provider="cosmere"
+      parentType="Homebrew"
+      publicationType="setting"
+      onFetchRequest={fetchList}
+      onFetchHomebrew={fetchHomebrew}
+      onBatchDestroy={batchDestroy}
+      onRemoveRequest={removeSettingRequest}
+      childrenComponent={ChildrenComponent}
+    />
+  );
+}