+++ app/javascript/applications/HomebrewsApp/pages/index.jsx
@@ -4,6 +4,7 @@ export * from './Content/Daggerheart';
 export * from './Content/Dnd2024';
 export * from './Content/Nimble';
 export * from './Content/Pathfinder2';
+export * from './Content/Cosmere';
 
 export * from './Content/SharedContent';
 export * from './Content/SharedBookContent';
@@ -37,6 +38,9 @@ export * from './Content/Dnd2024/Consumables';
 export * from './Content/Nimble/Ancestries';
 export * from './Content/Nimble/Books';
 export * from './Content/Nimble/Weapons';
+
+export * from './Content/Cosmere/Settings';
+export * from './Content/Cosmere/Books';
 
 export * from './Content/Pathfinder2/Books';
 export * from './Content/Pathfinder2/Weapons';